Split viessmann-helpers into focused lib/ modules (#76)
viessmann-helpers.js had grown to 573 LOC mixing four concerns: HTTP
transport, input validators, formatting, and Node-RED lifecycle/UI.
Adding cross-cutting features kept landing in the same file.

Split into:
- lib/client.js        - HTTP, transport, retry, token refresh
                         (already existed, unchanged)
- lib/validators.js    - validateConfigNode, validateInstallationId,
                         validateGatewaySerial, validateDeviceId,
                         validateFeature, validateCommand,
                         validateParams, viessmannRefSource,
                         validateViessmannRef
- lib/format.js        - extractErrorMessage, actionableStatusHint,
                         truncateForStatus
- lib/node-runtime.js  - initializeViessmannNode, createStatusUpdater,
                         setupDependentNode, surfaceUnexpectedError,
                         executeApiGet, executeApiPost

viessmann-helpers.js is now a 50-line barrel re-export so the existing
consumer-node files and tests keep importing from the same path. New
code can require from ./lib/* directly.

Closes #76
